+++ src/parser/enrichers/spell/MindSliver.mjs
@@ -1,0 +1,17 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../data/DDBEnricherData.mjs";
+
+export default class MindSliver extends DDBEnricherData {
+
+  get effects() {
+    return [
+      {
+        changes: [
+          DDBEnricherData.ChangeHelper.addChange("-1d4", 20, "system.bonuses.abilities.save"),
+        ],
+        daeSpecialDurations: ["isSave"],
+      },
+    ];
+  }
+
+}